Refactor for MVVM
